Create html page for "find shelters" feature
Added a way to navigate to this page from the main landing page. Currently it just contains an interactive map, so the shelter search functionality still needs to be added in.
For my own convenience, since index.html is quite large, I created a new html file so that I could start from a blank template for this feature.