Get working map location input, centering, and display results
User can now get autocompleted location suggestions when they enter in an address, and enter a search radius to stay within. The newly-added search button will then search for pet-related locations* around that location and display them as markers on the map. Clicking on the markers currently shows the shelter name and address--still need to add other info such as directions.

*Unfortunately, Geoapify doesn't have great data on animal shelters/humane societies. May swap to another API for searching for these specific locations.