FindShelters: use Overpass, change html name
Replace the name "test-map.html" with "findShelter.html", I only was using the old name as a placeholder.

Overpass has better shelter filtering so I'm using it for searching for animal shelters now. In some cases, it doesn't list the address with its results, so we supplement with reverse geocoding using geoapify to get an address in those cases.